fix: harden daemon observability boundary
