+++ .just/lint_boundaries.py
@@ -320,7 +320,7 @@ SCB_SINGLETON_ALLOWED_HOOK_CALLERS = {
     Path("crates/atm-daemon/src/composition.rs"),
 }
 SCB_OBSERVABILITY_ALLOWED_SRC_FILES = {
-    Path("crates/atm-daemon/src/daemon_runtime_observability.rs"),
+    Path("crates/atm-daemon/src/daemon_observability.rs"),
     Path("crates/atm-daemon/src/main.rs"),
 }
 SCB_OBSERVABILITY_DIRECT_PATTERNS = (
@@ -2766,7 +2766,7 @@ def collect_scb_observability_rule_violations(
                 continue
             violations.append(
                 BoundaryViolation(
-                    f"SCB-OBSERVABILITY-001 {rel_source}:{line_number} direct sc_observability_types ActionName/OutcomeLabel imports are forbidden outside daemon_runtime_observability.rs and main.rs",
+                    f"SCB-OBSERVABILITY-001 {rel_source}:{line_number} direct sc_observability_types ActionName/OutcomeLabel imports are forbidden outside daemon_observability.rs and main.rs",
                     "",
                 )
             )
