+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "crates/atm-storage",
     "crates/atm-storage-sqlserver-proof",
     "crates/atm-core",
+    "crates/atm-observability",
     "crates/atm-http-runtime",
     "crates/atm-herdr",
     "crates/atm-runtime",
@@ -63,6 +64,7 @@ syn = { version = "2", features = ["full", "visit"] }
 semver = { version = "1", features = ["serde"] }
 async-trait = "0.1"
 atm-core = { package = "agent-team-mail-core", path = "crates/atm-core", version = "1.4.6" }
+atm-observability = { path = "crates/atm-observability", version = "1.4.6" }
 atm-daemon-bootstrap = { path = "crates/atm-daemon-bootstrap", version = "1.4.6" }
 atm-daemon-client = { path = "crates/atm-daemon-client", version = "1.4.6" }
 atm-error = { path = "crates/atm-error", version = "1.4.6" }
