+++ boundaries/atm-daemon/daemon-observability.toml
@@ -1,0 +1,47 @@
+boundary_id = "BOUNDARY-ObservabilityPort-Daemon"
+owner_package = "atm-daemon"
+owner_crate_path = "atm_daemon"
+name = "DaemonObservability"
+
+[public]
+trait = "ObservabilityPort"
+notes = "DaemonObservability is the permit-listed retained-log implementation of the sealed ATM observability port for the active Tokio/Axum daemon entrypoint."
+
+[implementation]
+type = "DaemonObservability"
+module = "atm_daemon::daemon_observability"
+visibility = "public"
+constructor = "public"
+
+[composition]
+roots = ["atm_daemon::main::run"]
+
+[ownership]
+io_owns = ["retained_log_bootstrap", "retained_log_emit", "retained_log_health_projection"]
+io_forbidden = ["daemon_request_dispatch", "daemon_transport", "direct_sqlite_io", "daemon_runtime_bootstrap"]
+
+[dependencies]
+allowed_dependents = []
+allowed_dependencies = ["atm-core", "atm-daemon-bootstrap", "atm-observability", "chrono", "sc-observability", "sc-observability-types", "serde_json", "serial_test", "tempfile", "tokio", "tracing"]
+forbidden_edges = ["atm-daemon -> atm-storage-rusqlite", "atm-daemon -> atm-runtime", "atm-daemon -> atm-http-runtime"]
+
+[references]
+scope = "outside_owner_crate"
+forbidden = ["legacy_daemon_server", "legacy_dispatcher"]
+
+[contracts]
+request_types = ["CommandEvent", "AtmLogQuery"]
+response_types = ["AtmLogSnapshot", "LogTailSession", "AtmObservabilityHealth"]
+error_types = ["AtmError"]
+
+[testing]
+allowed_test_double_paths = []
+forbidden_test_bypasses = ["legacy_daemon_server", "legacy_dispatcher"]
+
+[enforcement]
+lint_rules = ["SCB-OBSERVABILITY-001", "LINT-BOUNDARY-DAEMON-OBSERVABILITY-DEPENDENCIES"]
+review_gates = ["sealed_impl_allowlisted", "tokio_axum_entrypoint_only", "no_legacy_daemon_composition"]
+
+[status]
+state = "active"
+notes = ["The concrete adapter is source-visible to lint_boundaries.py; no bin_support include shim may hide its sealed implementation.", "The binary delegates runtime lifecycle to atm-daemon-bootstrap and the Tokio/Axum atm-http-runtime path."]
